hw2/requirements.txt
requests
langchain
langchain_google_genai
wikipedia
langchain-experimental
langchainhub
e2b
numexpr
python-dotenv